refactor(mcp): extract parsing, render, compare into tools/ modules
Begin splitting the 12.3k-line server.rs. Each domain moves to a
src/mcp/tools/<domain>.rs as an additional `impl McpServer` block, so the
dispatch and tests call them unchanged via Self::; cross-module helpers
(parse_*, validate_path, render/compare fns) become pub(crate).

- tools/parsing.rs: 14 JSON->primitive parse_* helpers (~770 lines)
- tools/render.rs: ASCII footprint/symbol rendering (~730 lines)
- tools/compare.rs: compare_components + compare_* (~900 lines)

server.rs: 12299 -> ~9890 lines so far. Build/clippy/tests green.
