+++ src/mcp/mod.rs
@@ -29,6 +29,7 @@
 pub mod protocol;
 pub mod server;
 mod tool_definitions;
+mod tools;
 pub mod transport;
 
 pub use protocol::{JsonRpcError, JsonRpcRequest, JsonRpcResponse, MCP_PROTOCOL_VERSION};
