+++ src/mcp/server.rs
@@ -351,7 +351,7 @@ impl McpServer {
     /// Validates that a path is within one of the allowed paths.
     ///
     /// Returns `Ok(())` if the path is allowed, or an error message if not.
-    fn validate_path(&self, filepath: &str) -> Result<(), String> {
+    pub(crate) fn validate_path(&self, filepath: &str) -> Result<(), String> {
         use std::path::Path;
 
         // Fail closed: with no configured allowed paths, deny everything rather
@@ -2230,779 +2230,6 @@ impl McpServer {
         }
 
         Ok(())
-    }
-
-    // ==================== Primitive Parsing Helpers ====================
-
-    /// Parses a pad from JSON.
-    #[allow(clippy::too_many_lines)] // Pad has many fields requiring individual parsing
-    fn parse_pad(json: &Value) -> Result<crate::altium::pcblib::Pad, String> {
-        use crate::altium::pcblib::{Layer, Pad, PadShape, PadStackMode, PcbFlags};
-
-        let designator = json
-            .get("designator")
-            .and_then(Value::as_str)
-            .ok_or("Pad missing required field 'designator'")?;
-
-        // Validate designator is not empty
-        if designator.trim().is_empty() {
-            return Err("Pad designator cannot be empty".to_string());
-        }
-
-        let x = json
-            .get("x")
-            .and_then(Value::as_f64)
-            .ok_or("Pad missing required field 'x'")?;
-        let y = json
-            .get("y")
-            .and_then(Value::as_f64)
-            .ok_or("Pad missing required field 'y'")?;
-        let width = json
-            .get("width")
-            .and_then(Value::as_f64)
-            .ok_or("Pad missing required field 'width'")?;
-        let height = json
-            .get("height")
-            .and_then(Value::as_f64)
-            .ok_or("Pad missing required field 'height'")?;
-
-        // Validate pad dimensions are positive
-        if width <= 0.0 {
-            return Err(format!(
-                "Pad '{designator}' has invalid width {width}. Width must be greater than 0."
-            ));
-        }
-        if height <= 0.0 {
-            return Err(format!(
-                "Pad '{designator}' has invalid height {height}. Height must be greater than 0."
-            ));
-        }
-
-        let shape_str = json
-            .get("shape")
-            .and_then(Value::as_str)
-            .unwrap_or("rounded_rectangle");
-        let shape = match shape_str {
-            "rectangle" => PadShape::Rectangle,
-            "round" | "circle" => PadShape::Round,
-            "oval" => PadShape::Oval,
-            "rounded_rectangle" => PadShape::RoundedRectangle,
-            _ => {
-                return Err(format!(
-                    "Pad '{designator}' has invalid shape '{shape_str}'. \
-                     Valid shapes are: rectangle, round, circle, oval, rounded_rectangle"
-                ))
-            }
-        };
-
-        // Parse hole_size first to determine default layer
-        let hole_size = json.get("hole_size").and_then(Value::as_f64);
-        let is_smd = hole_size.map_or(true, |h| h <= 0.0); // SMD if no hole or hole size <= 0
-
-        let layer_str = json.get("layer").and_then(Value::as_str);
-        let layer = match layer_str {
-            Some(s) => Layer::parse(s).ok_or_else(|| {
-                format!(
-                    "Pad '{designator}' has invalid layer '{s}'. \
-                     Valid layers include: Top Layer, Bottom Layer, Multi-Layer, Top Overlay, etc."
-                )
-            })?,
-            // SMD pads default to Top Layer, through-hole pads default to Multi-Layer
-            None => {
-                if is_smd {
-                    Layer::TopLayer
-                } else {
-                    Layer::MultiLayer
-                }
-            }
-        };
-        let rotation = json.get("rotation").and_then(Value::as_f64).unwrap_or(0.0);
-
-        // Parse optional hole shape
-        let hole_shape = json
-            .get("hole_shape")
-            .and_then(Value::as_str)
-            .map(|s| match s.to_lowercase().as_str() {
-                "square" => crate::altium::pcblib::HoleShape::Square,
-                "slot" => crate::altium::pcblib::HoleShape::Slot,
-                _ => crate::altium::pcblib::HoleShape::Round,
-            })
-            .unwrap_or_default();
-
-        // Parse optional mask expansion values
-        let paste_mask_expansion = json.get("paste_mask_expansion").and_then(Value::as_f64);
-        let solder_mask_expansion = json.get("solder_mask_expansion").and_then(Value::as_f64);
-        let paste_mask_expansion_manual = json
-            .get("paste_mask_expansion_manual")
-            .and_then(Value::as_bool)
-            .unwrap_or(false);
-        let solder_mask_expansion_manual = json
-            .get("solder_mask_expansion_manual")
-            .and_then(Value::as_bool)
-            .unwrap_or(false);
-
-        // Parse optional corner radius
-        let corner_radius_percent = json
-            .get("corner_radius_percent")
-            .and_then(Value::as_u64)
-            .and_then(|v| u8::try_from(v).ok())
-            .filter(|&v| v <= 100);
-
-        Ok(Pad {
-            designator: designator.to_string(),
-            x,
-            y,
-            width,
-            height,
-            shape,
-            layer,
-            hole_size,
-            hole_shape,
-            rotation,
-            paste_mask_expansion,
-            solder_mask_expansion,
-            paste_mask_expansion_manual,
-            solder_mask_expansion_manual,
-            corner_radius_percent,
-            stack_mode: PadStackMode::Simple,
-            per_layer_sizes: None,
-            per_layer_shapes: None,
-            per_layer_corner_radii: None,
-            per_layer_offsets: None,
-            flags: PcbFlags::empty(),
-            unique_id: None,
-        })
-    }
-
-    /// Parses a track from JSON.
-    fn parse_track(json: &Value) -> Result<crate::altium::pcblib::Track, String> {
-        use crate::altium::pcblib::{Layer, Track};
-
-        let x1 = json
-            .get("x1")
-            .and_then(Value::as_f64)
-            .ok_or("Track missing required field 'x1'")?;
-        let y1 = json
-            .get("y1")
-            .and_then(Value::as_f64)
-            .ok_or("Track missing required field 'y1'")?;
-        let x2 = json
-            .get("x2")
-            .and_then(Value::as_f64)
-            .ok_or("Track missing required field 'x2'")?;
-        let y2 = json
-            .get("y2")
-            .and_then(Value::as_f64)
-            .ok_or("Track missing required field 'y2'")?;
-        let width = json
-            .get("width")
-            .and_then(Value::as_f64)
-            .ok_or("Track missing required field 'width'")?;
-
-        let layer_str = json.get("layer").and_then(Value::as_str);
-        let layer = match layer_str {
-            Some(s) => Layer::parse(s).ok_or_else(|| {
-                format!(
-                    "Track has invalid layer '{s}'. \
-                     Valid layers include: Top Layer, Bottom Layer, Top Overlay, Top Assembly, etc."
-                )
-            })?,
-            None => Layer::TopOverlay, // Default for tracks is Top Overlay
-        };
-
-        Ok(Track::new(x1, y1, x2, y2, width, layer))
-    }
-
-    /// Parses an arc from JSON.
-    fn parse_arc(json: &Value) -> Result<crate::altium::pcblib::Arc, String> {
-        use crate::altium::pcblib::{Arc, Layer, PcbFlags};
-
-        let x = json
-            .get("x")
-            .and_then(Value::as_f64)
-            .ok_or("Arc missing required field 'x'")?;
-        let y = json
-            .get("y")
-            .and_then(Value::as_f64)
-            .ok_or("Arc missing required field 'y'")?;
-        let radius = json
-            .get("radius")
-            .and_then(Value::as_f64)
-            .ok_or("Arc missing required field 'radius'")?;
-        let start_angle = json
-            .get("start_angle")
-            .and_then(Value::as_f64)
-            .ok_or("Arc missing required field 'start_angle'")?;
-        let end_angle = json
-            .get("end_angle")
-            .and_then(Value::as_f64)
-            .ok_or("Arc missing required field 'end_angle'")?;
-        let width = json
-            .get("width")
-            .and_then(Value::as_f64)
-            .ok_or("Arc missing required field 'width'")?;
-
-        let layer_str = json.get("layer").and_then(Value::as_str);
-        let layer = match layer_str {
-            Some(s) => Layer::parse(s).ok_or_else(|| {
-                format!(
-                    "Arc has invalid layer '{s}'. \
-                     Valid layers include: Top Layer, Bottom Layer, Top Overlay, Top Assembly, etc."
-                )
-            })?,
-            None => Layer::TopOverlay, // Default for arcs is Top Overlay
-        };
-
-        Ok(Arc {
-            x,
-            y,
-            radius,
-            start_angle,
-            end_angle,
-            width,
-            layer,
-            flags: PcbFlags::empty(),
-            unique_id: None,
-        })
-    }
-
-    /// Parses a region from JSON.
-    fn parse_region(json: &Value) -> Option<crate::altium::pcblib::Region> {
-        use crate::altium::pcblib::{Layer, PcbFlags, Region, Vertex};
-
-        let vertices_json = json.get("vertices").and_then(Value::as_array)?;
-        let layer = json
-            .get("layer")
-            .and_then(Value::as_str)
-            .and_then(Layer::parse)
-            .unwrap_or(Layer::Mechanical15);
-
-        let vertices: Vec<Vertex> = vertices_json
-            .iter()
-            .filter_map(|v| {
-                let x = v.get("x").and_then(Value::as_f64)?;
-                let y = v.get("y").and_then(Value::as_f64)?;
-                Some(Vertex { x, y })
-            })
-            .collect();
-
-        if vertices.len() < 3 {
-            return None; // Need at least 3 vertices for a polygon
-        }
-
-        Some(Region {
-            vertices,
-            layer,
-            flags: PcbFlags::empty(),
-            unique_id: None,
-        })
-    }
-
-    /// Parses text from JSON.
-    fn parse_text(json: &Value) -> Option<crate::altium::pcblib::Text> {
-        use crate::altium::pcblib::{Layer, PcbFlags, Text, TextJustification, TextKind};
-
-        let x = json.get("x").and_then(Value::as_f64)?;
-        let y = json.get("y").and_then(Value::as_f64)?;
-        let text = json.get("text").and_then(Value::as_str)?;
-        let height = json.get("height").and_then(Value::as_f64)?;
-        let layer = json
-            .get("layer")
-            .and_then(Value::as_str)
-            .and_then(Layer::parse)
-            .unwrap_or(Layer::TopOverlay);
-        let rotation = json.get("rotation").and_then(Value::as_f64).unwrap_or(0.0);
-
-        Some(Text {
-            x,
-            y,
-            text: text.to_string(),
-            height,
-            layer,
-            rotation,
-            kind: TextKind::Stroke,
-            stroke_font: None,
-            justification: TextJustification::MiddleCenter,
-            flags: PcbFlags::empty(),
-            unique_id: None,
-        })
-    }
-
-    // ==================== SchLib Primitive Parsing Helpers ====================
-
-    /// Parses a schematic pin from JSON.
-    #[allow(clippy::cast_possible_truncation)]
-    #[allow(clippy::too_many_lines)] // Pin parsing with symbol attributes requires many lines
-    fn parse_schlib_pin(json: &Value) -> Option<crate::altium::schlib::Pin> {
-        use crate::altium::schlib::{Pin, PinElectricalType, PinOrientation, PinSymbol};
-
-        let designator = json.get("designator").and_then(Value::as_str)?;
-        let name = json
-            .get("name")
-            .and_then(Value::as_str)
-            .unwrap_or(designator);
-        let x = json.get("x").and_then(Value::as_i64)? as i32;
-        let y = json.get("y").and_then(Value::as_i64)? as i32;
-        let length = json.get("length").and_then(Value::as_i64).unwrap_or(10) as i32;
-
-        let orientation =
-            json.get("orientation")
-                .and_then(Value::as_str)
-                .map_or(PinOrientation::Right, |s| match s.to_lowercase().as_str() {
-                    "left" => PinOrientation::Left,
-                    "up" => PinOrientation::Up,
-                    "down" => PinOrientation::Down,
-                    _ => PinOrientation::Right,
-                });
-
-        let electrical_type = json.get("electrical_type").and_then(Value::as_str).map_or(
-            PinElectricalType::Passive,
-            |s| match s.to_lowercase().as_str() {
-                "input" => PinElectricalType::Input,
-                "output" => PinElectricalType::Output,
-                "bidirectional" | "io" | "input_output" => PinElectricalType::Bidirectional,
-                "power" => PinElectricalType::Power,
-                "open_collector" => PinElectricalType::OpenCollector,
-                "open_emitter" => PinElectricalType::OpenEmitter,
-                "hiz" | "hi_z" | "tristate" => PinElectricalType::HiZ,
-                _ => PinElectricalType::Passive,
-            },
-        );
-
-        let hidden = json.get("hidden").and_then(Value::as_bool).unwrap_or(false);
-        let show_name = json
-            .get("show_name")
-            .and_then(Value::as_bool)
-            .unwrap_or(true);
-        let show_designator = json
-            .get("show_designator")
-            .and_then(Value::as_bool)
-            .unwrap_or(true);
-        let owner_part_id = json
-            .get("owner_part_id")
-            .and_then(Value::as_i64)
-            .unwrap_or(1) as i32;
-
-        // Helper to parse PinSymbol from string
-        let parse_pin_symbol = |s: &str| -> PinSymbol {
-            match s.to_lowercase().replace(['-', '_'], "").as_str() {
-                "dot" | "invert" | "inversion" => PinSymbol::Dot,
-                "clock" | "clk" => PinSymbol::Clock,
-                "activelowinput" | "lowinput" => PinSymbol::ActiveLowInput,
-                "activelowoutput" | "lowoutput" => PinSymbol::ActiveLowOutput,
-                "rightleftsignalflow" | "rightleft" => PinSymbol::RightLeftSignalFlow,
-                "leftrrightsignalflow" | "leftright" => PinSymbol::LeftRightSignalFlow,
-                "bidirectionalsignalflow" | "bidirectional" => PinSymbol::BidirectionalSignalFlow,
-                "analogsignalin" | "analog" => PinSymbol::AnalogSignalIn,
-                "digitalsignalin" | "digital" => PinSymbol::DigitalSignalIn,
-                "notlogicconnection" | "notlogic" => PinSymbol::NotLogicConnection,
-                "postponedoutput" | "postponed" => PinSymbol::PostponedOutput,
-                "opencollector" => PinSymbol::OpenCollector,
-                "opencollectorpullup" => PinSymbol::OpenCollectorPullUp,
-                "openemitter" => PinSymbol::OpenEmitter,
-                "openemitterpullup" => PinSymbol::OpenEmitterPullUp,
-                "openoutput" => PinSymbol::OpenOutput,
-                "hiz" | "highimpedance" => PinSymbol::HiZ,
-                "highcurrent" => PinSymbol::HighCurrent,
-                "pulse" => PinSymbol::Pulse,
-                "schmitt" | "schmitttrigger" => PinSymbol::Schmitt,
-                "shiftleft" => PinSymbol::ShiftLeft,
-                _ => PinSymbol::None, // "none" and unknown values
-            }
-        };
-
-        let symbol_inner_edge = json
-            .get("symbol_inner_edge")
-            .and_then(Value::as_str)
-            .map_or(PinSymbol::None, parse_pin_symbol);
-        let symbol_outer_edge = json
-            .get("symbol_outer_edge")
-            .and_then(Value::as_str)
-            .map_or(PinSymbol::None, parse_pin_symbol);
-        let symbol_inside = json
-            .get("symbol_inside")
-            .and_then(Value::as_str)
-            .map_or(PinSymbol::None, parse_pin_symbol);
-        let symbol_outside = json
-            .get("symbol_outside")
-            .and_then(Value::as_str)
-            .map_or(PinSymbol::None, parse_pin_symbol);
-
-        Some(Pin {
-            name: name.to_string(),
-            designator: designator.to_string(),
-            x,
-            y,
-            length,
-            orientation,
-            electrical_type,
-            hidden,
-            show_name,
-            show_designator,
-            description: String::new(),
-            owner_part_id,
-            colour: 0,
-            graphically_locked: false,
-            symbol_inner_edge,
-            symbol_outer_edge,
-            symbol_inside,
-            symbol_outside,
-        })
-    }
-
-    /// Parses a schematic rectangle from JSON.
-    #[allow(clippy::cast_possible_truncation)]
-    fn parse_schlib_rectangle(json: &Value) -> Option<crate::altium::schlib::Rectangle> {
-        use crate::altium::schlib::Rectangle;
-
-        let x1 = json.get("x1").and_then(Value::as_i64)? as i32;
-        let y1 = json.get("y1").and_then(Value::as_i64)? as i32;
-        let x2 = json.get("x2").and_then(Value::as_i64)? as i32;
-        let y2 = json.get("y2").and_then(Value::as_i64)? as i32;
-
-        let line_width = json.get("line_width").and_then(Value::as_u64).unwrap_or(1) as u8;
-        let line_color = json
-            .get("line_color")
-            .and_then(Value::as_u64)
-            .unwrap_or(0x00_00_80) as u32;
-        let fill_color = json
-            .get("fill_color")
-            .and_then(Value::as_u64)
-            .unwrap_or(0xFF_FF_B0) as u32;
-        let filled = json.get("filled").and_then(Value::as_bool).unwrap_or(true);
-        let owner_part_id = json
-            .get("owner_part_id")
-            .and_then(Value::as_i64)
-            .unwrap_or(1) as i32;
-
-        Some(Rectangle {
-            x1,
-            y1,
-            x2,
-            y2,
-            line_width,
-            line_color,
-            fill_color,
-            filled,
-            transparent: false,
-            owner_part_id,
-        })
-    }
-
-    /// Parses a schematic line from JSON.
-    #[allow(clippy::cast_possible_truncation)]
-    fn parse_schlib_line(json: &Value) -> Option<crate::altium::schlib::Line> {
-        use crate::altium::schlib::Line;
-
-        let x1 = json.get("x1").and_then(Value::as_i64)? as i32;
-        let y1 = json.get("y1").and_then(Value::as_i64)? as i32;
-        let x2 = json.get("x2").and_then(Value::as_i64)? as i32;
-        let y2 = json.get("y2").and_then(Value::as_i64)? as i32;
-
-        let line_width = json.get("line_width").and_then(Value::as_u64).unwrap_or(1) as u8;
-        let color = json
-            .get("color")
-            .and_then(Value::as_u64)
-            .unwrap_or(0x00_00_80) as u32;
-        let owner_part_id = json
-            .get("owner_part_id")
-            .and_then(Value::as_i64)
-            .unwrap_or(1) as i32;
-
-        Some(Line {
-            x1,
-            y1,
-            x2,
-            y2,
-            line_width,
-            color,
-            owner_part_id,
-        })
-    }
-
-    /// Parses a schematic parameter from JSON.
-    #[allow(clippy::cast_possible_truncation)]
-    fn parse_schlib_parameter(json: &Value) -> Option<crate::altium::schlib::Parameter> {
-        use crate::altium::schlib::Parameter;
-
-        let name = json.get("name").and_then(Value::as_str)?;
-        let value = json
-            .get("value")
-            .and_then(Value::as_str)
-            .unwrap_or("*")
-            .to_string();
-
-        let x = json.get("x").and_then(Value::as_i64).unwrap_or(0) as i32;
-        let y = json.get("y").and_then(Value::as_i64).unwrap_or(0) as i32;
-        let font_id = json.get("font_id").and_then(Value::as_u64).unwrap_or(1) as u8;
-        let color = json
-            .get("color")
-            .and_then(Value::as_u64)
-            .unwrap_or(0x80_00_00) as u32;
-        let hidden = json.get("hidden").and_then(Value::as_bool).unwrap_or(false);
-        let owner_part_id = json
-            .get("owner_part_id")
-            .and_then(Value::as_i64)
-            .unwrap_or(1) as i32;
-
-        Some(Parameter {
-            name: name.to_string(),
-            value,
-            x,
-            y,
-            font_id,
-            color,
-            hidden,
-            read_only_state: 0,
-            param_type: 0,
-            owner_part_id,
-        })
-    }
-
-    /// Parses a schematic polyline from JSON.
-    /// Accepts both "points" and "vertices" field names for the coordinate array.
-    #[allow(clippy::cast_possible_truncation)]
-    fn parse_schlib_polyline(json: &Value) -> Option<crate::altium::schlib::Polyline> {
-        use crate::altium::schlib::Polyline;
-
-        // Accept both "points" and "vertices" for flexibility
-        let points_json = json
-            .get("points")
-            .or_else(|| json.get("vertices"))
-            .and_then(Value::as_array)?;
-        let points: Vec<(i32, i32)> = points_json
-            .iter()
-            .filter_map(|p| {
-                let x = p.get("x").and_then(Value::as_i64)? as i32;
-                let y = p.get("y").and_then(Value::as_i64)? as i32;
-                Some((x, y))
-            })
-            .collect();
-
-        if points.len() < 2 {
-            return None; // Need at least 2 points for a polyline
-        }
-
-        let line_width = json.get("line_width").and_then(Value::as_u64).unwrap_or(1) as u8;
-        let color = json
-            .get("color")
-            .and_then(Value::as_u64)
-            .unwrap_or(0x00_00_80) as u32;
-        let owner_part_id = json
-            .get("owner_part_id")
-            .and_then(Value::as_i64)
-            .unwrap_or(1) as i32;
-
-        Some(Polyline {
-            points,
-            line_width,
-            color,
-            line_style: 0,
-            start_line_shape: 0,
-            end_line_shape: 0,
-            line_shape_size: 0,
-            owner_part_id,
-        })
-    }
-
-    /// Parses a schematic arc from JSON.
-    #[allow(clippy::cast_possible_truncation)]
-    fn parse_schlib_arc(json: &Value) -> Option<crate::altium::schlib::Arc> {
-        use crate::altium::schlib::Arc;
-
-        let x = json.get("x").and_then(Value::as_i64)? as i32;
-        let y = json.get("y").and_then(Value::as_i64)? as i32;
-        let radius = json.get("radius").and_then(Value::as_i64)? as i32;
-        let start_angle = json
-            .get("start_angle")
-            .and_then(Value::as_f64)
-            .unwrap_or(0.0);
-        let end_angle = json
-            .get("end_angle")
-            .and_then(Value::as_f64)
-            .unwrap_or(360.0);
-        let line_width = json.get("line_width").and_then(Value::as_u64).unwrap_or(1) as u8;
-        let color = json
-            .get("color")
-            .and_then(Value::as_u64)
-            .unwrap_or(0x00_00_80) as u32;
-        let owner_part_id = json
-            .get("owner_part_id")
-            .and_then(Value::as_i64)
-            .unwrap_or(1) as i32;
-
-        Some(Arc {
-            x,
-            y,
-            radius,
-            start_angle,
-            end_angle,
-            line_width,
-            color,
-            owner_part_id,
-        })
-    }
-
-    /// Parses a schematic ellipse from JSON.
-    #[allow(clippy::cast_possible_truncation)]
-    fn parse_schlib_ellipse(json: &Value) -> Option<crate::altium::schlib::Ellipse> {
-        use crate::altium::schlib::Ellipse;
-
-        let x = json.get("x").and_then(Value::as_i64)? as i32;
-        let y = json.get("y").and_then(Value::as_i64)? as i32;
-        let radius_x = json.get("radius_x").and_then(Value::as_i64)? as i32;
-        let radius_y = json.get("radius_y").and_then(Value::as_i64)? as i32;
-
-        let line_width = json.get("line_width").and_then(Value::as_u64).unwrap_or(1) as u8;
-        let line_color = json
-            .get("line_color")
-            .and_then(Value::as_u64)
-            .unwrap_or(0x00_00_80) as u32;
-        let fill_color = json
-            .get("fill_color")
-            .and_then(Value::as_u64)
-            .unwrap_or(0xFF_FF_B0) as u32;
-        let filled = json.get("filled").and_then(Value::as_bool).unwrap_or(true);
-        let owner_part_id = json
-            .get("owner_part_id")
-            .and_then(Value::as_i64)
-            .unwrap_or(1) as i32;
-
-        Some(Ellipse {
-            x,
-            y,
-            radius_x,
-            radius_y,
-            line_width,
-            line_color,
-            fill_color,
-            filled,
-            owner_part_id,
-        })
-    }
-
-    /// Parses a schematic label from JSON.
-    #[allow(clippy::cast_possible_truncation)]
-    fn parse_schlib_label(json: &Value) -> Option<crate::altium::schlib::Label> {
-        use crate::altium::schlib::{Label, TextJustification};
-
-        let x = json.get("x").and_then(Value::as_i64)? as i32;
-        let y = json.get("y").and_then(Value::as_i64)? as i32;
-        let text = json.get("text").and_then(Value::as_str)?.to_string();
-
-        let font_id = json.get("font_id").and_then(Value::as_u64).unwrap_or(1) as u8;
-        let color = json
-            .get("color")
-            .and_then(Value::as_u64)
-            .unwrap_or(0x00_00_80) as u32;
-        let rotation = json.get("rotation").and_then(Value::as_f64).unwrap_or(0.0);
-        let is_mirrored = json
-            .get("is_mirrored")
-            .and_then(Value::as_bool)
-            .unwrap_or(false);
-        let is_hidden = json
-            .get("is_hidden")
-            .or_else(|| json.get("hidden"))
-            .and_then(Value::as_bool)
-            .unwrap_or(false);
-        let owner_part_id = json
-            .get("owner_part_id")
-            .and_then(Value::as_i64)
-            .unwrap_or(1) as i32;
-
-        let justification = json.get("justification").and_then(Value::as_str).map_or(
-            TextJustification::BottomLeft,
-            |s| {
-                match s.to_lowercase().replace(['-', '_'], "").as_str() {
-                    "bottomcenter" | "bottomcentre" => TextJustification::BottomCenter,
-                    "bottomright" => TextJustification::BottomRight,
-                    "middleleft" | "centerleft" | "centreleft" => TextJustification::MiddleLeft,
-                    "middlecenter" | "middlecentre" | "center" | "centre" => {
-                        TextJustification::MiddleCenter
-                    }
-                    "middleright" | "centerright" | "centreright" => TextJustification::MiddleRight,
-                    "topleft" => TextJustification::TopLeft,
-                    "topcenter" | "topcentre" => TextJustification::TopCenter,
-                    "topright" => TextJustification::TopRight,
-                    _ => TextJustification::BottomLeft, // "bottomleft" and unknown values
-                }
-            },
-        );
-
-        Some(Label {
-            x,
-            y,
-            text,
-            font_id,
-            color,
-            justification,
-            rotation,
-            is_mirrored,
-            is_hidden,
-            owner_part_id,
-        })
-    }
-
-    /// Parses a schematic text annotation from JSON.
-    #[allow(clippy::cast_possible_truncation)]
-    fn parse_schlib_text(json: &Value) -> Option<crate::altium::schlib::Text> {
-        use crate::altium::schlib::{Text, TextJustification};
-
-        let x = json.get("x").and_then(Value::as_i64)? as i32;
-        let y = json.get("y").and_then(Value::as_i64)? as i32;
-        let text = json.get("text").and_then(Value::as_str)?.to_string();
-
-        let font_id = json.get("font_id").and_then(Value::as_u64).unwrap_or(1) as u8;
-        let color = json
-            .get("color")
-            .and_then(Value::as_u64)
-            .unwrap_or(0x00_00_80) as u32;
-        let rotation = json.get("rotation").and_then(Value::as_f64).unwrap_or(0.0);
-        let is_mirrored = json
-            .get("is_mirrored")
-            .and_then(Value::as_bool)
-            .unwrap_or(false);
-        let is_hidden = json
-            .get("is_hidden")
-            .or_else(|| json.get("hidden"))
-            .and_then(Value::as_bool)
-            .unwrap_or(false);
-        let owner_part_id = json
-            .get("owner_part_id")
-            .and_then(Value::as_i64)
-            .unwrap_or(1) as i32;
-
-        let justification = json.get("justification").and_then(Value::as_str).map_or(
-            TextJustification::BottomLeft,
-            |s| {
-                match s.to_lowercase().replace(['-', '_'], "").as_str() {
-                    "bottomcenter" | "bottomcentre" => TextJustification::BottomCenter,
-                    "bottomright" => TextJustification::BottomRight,
-                    "middleleft" | "centerleft" | "centreleft" => TextJustification::MiddleLeft,
-                    "middlecenter" | "middlecentre" | "center" | "centre" => {
-                        TextJustification::MiddleCenter
-                    }
-                    "middleright" | "centerright" | "centreright" => TextJustification::MiddleRight,
-                    "topleft" => TextJustification::TopLeft,
-                    "topcenter" | "topcentre" => TextJustification::TopCenter,
-                    "topright" => TextJustification::TopRight,
-                    _ => TextJustification::BottomLeft, // "bottomleft" and unknown values
-                }
-            },
-        );
-
-        Some(Text {
-            x,
-            y,
-            text,
-            font_id,
-            color,
-            justification,
-            rotation,
-            is_mirrored,
-            is_hidden,
-            owner_part_id,
-        })
     }
 
     // ==================== Library Management Tools ====================
@@ -5074,909 +4301,6 @@ impl McpServer {
         });
 
         ToolCallResult::text(serde_json::to_string_pretty(&result).unwrap())
-    }
-
-    /// Compares two specific components in detail.
-    fn call_compare_components(&self, arguments: &Value) -> ToolCallResult {
-        let Some(filepath_a) = arguments.get("filepath_a").and_then(Value::as_str) else {
-            return ToolCallResult::error("Missing required parameter: filepath_a");
-        };
-        let Some(component_a) = arguments.get("component_a").and_then(Value::as_str) else {
-            return ToolCallResult::error("Missing required parameter: component_a");
-        };
-        let Some(filepath_b) = arguments.get("filepath_b").and_then(Value::as_str) else {
-            return ToolCallResult::error("Missing required parameter: filepath_b");
-        };
-        let Some(component_b) = arguments.get("component_b").and_then(Value::as_str) else {
-            return ToolCallResult::error("Missing required parameter: component_b");
-        };
-
-        // Validate paths
-        if let Err(e) = self.validate_path(filepath_a) {
-            return ToolCallResult::error(e);
-        }
-        if let Err(e) = self.validate_path(filepath_b) {
-            return ToolCallResult::error(e);
-        }
-
-        let include_geometry = arguments
-            .get("include_geometry")
-            .and_then(Value::as_bool)
-            .unwrap_or(true);
-
-        let tolerance = arguments
-            .get("tolerance")
-            .and_then(Value::as_f64)
-            .unwrap_or(0.001);
-
-        // Determine file types
-        let path_a = std::path::Path::new(filepath_a);
-        let path_b = std::path::Path::new(filepath_b);
-
-        let ext_a = path_a
-            .extension()
-            .and_then(|e| e.to_str())
-            .map(str::to_lowercase);
-        let ext_b = path_b
-            .extension()
-            .and_then(|e| e.to_str())
-            .map(str::to_lowercase);
-
-        // Ensure both files are the same type
-        if ext_a != ext_b {
-            return ToolCallResult::error(format!(
-                "File types must match. Got '{}' and '{}'",
-                ext_a.as_deref().unwrap_or("unknown"),
-                ext_b.as_deref().unwrap_or("unknown")
-            ));
-        }
-
-        match ext_a.as_deref() {
-            Some("pcblib") => Self::compare_footprints(
-                filepath_a,
-                component_a,
-                filepath_b,
-                component_b,
-                include_geometry,
-                tolerance,
-            ),
-            Some("schlib") => Self::compare_symbols(
-                filepath_a,
-                component_a,
-                filepath_b,
-                component_b,
-                include_geometry,
-            ),
-            _ => ToolCallResult::error("Unknown file type. Expected .PcbLib or .SchLib extension."),
-        }
-    }
-
-    /// Compares two footprints in detail.
-    #[allow(clippy::too_many_lines)]
-    fn compare_footprints(
-        filepath_a: &str,
-        name_a: &str,
-        filepath_b: &str,
-        name_b: &str,
-        include_geometry: bool,
-        tolerance: f64,
-    ) -> ToolCallResult {
-        use crate::altium::PcbLib;
-
-        // Read libraries
-        let lib_a = match PcbLib::open(filepath_a) {
-            Ok(lib) => lib,
-            Err(e) => return ToolCallResult::error(format!("Failed to read '{filepath_a}': {e}")),
-        };
-        let lib_b = match PcbLib::open(filepath_b) {
-            Ok(lib) => lib,
-            Err(e) => return ToolCallResult::error(format!("Failed to read '{filepath_b}': {e}")),
-        };
-
-        // Get components
-        let Some(fp_a) = lib_a.get(name_a) else {
-            return ToolCallResult::error(format!(
-                "Component '{name_a}' not found in '{filepath_a}'"
-            ));
-        };
-        let Some(fp_b) = lib_b.get(name_b) else {
-            return ToolCallResult::error(format!(
-                "Component '{name_b}' not found in '{filepath_b}'"
-            ));
-        };
-
-        let mut differences: Vec<Value> = Vec::new();
-
-        // Compare description
-        if fp_a.description != fp_b.description {
-            differences.push(json!({
-                "field": "description",
-                "component_a": fp_a.description,
-                "component_b": fp_b.description
-            }));
-        }
-
-        // Compare pad counts
-        if fp_a.pads.len() != fp_b.pads.len() {
-            differences.push(json!({
-                "field": "pad_count",
-                "component_a": fp_a.pads.len(),
-                "component_b": fp_b.pads.len()
-            }));
-        }
-
-        // Compare pads in detail
-        if include_geometry {
-            let pad_diffs = Self::compare_pads(&fp_a.pads, &fp_b.pads, tolerance);
-            if !pad_diffs.is_empty() {
-                differences.push(json!({
-                    "field": "pads",
-                    "differences": pad_diffs
-                }));
-            }
-        }
-
-        // Compare track counts
-        if fp_a.tracks.len() != fp_b.tracks.len() {
-            differences.push(json!({
-                "field": "track_count",
-                "component_a": fp_a.tracks.len(),
-                "component_b": fp_b.tracks.len()
-            }));
-        }
-
-        // Compare tracks in detail
-        if include_geometry {
-            let track_diffs = Self::compare_tracks(&fp_a.tracks, &fp_b.tracks, tolerance);
-            if !track_diffs.is_empty() {
-                differences.push(json!({
-                    "field": "tracks",
-                    "differences": track_diffs
-                }));
-            }
-        }
-
-        // Compare arc counts
-        if fp_a.arcs.len() != fp_b.arcs.len() {
-            differences.push(json!({
-                "field": "arc_count",
-                "component_a": fp_a.arcs.len(),
-                "component_b": fp_b.arcs.len()
-            }));
-        }
-
-        // Compare arcs in detail
-        if include_geometry {
-            let arc_diffs = Self::compare_pcb_arcs(&fp_a.arcs, &fp_b.arcs, tolerance);
-            if !arc_diffs.is_empty() {
-                differences.push(json!({
-                    "field": "arcs",
-                    "differences": arc_diffs
-                }));
-            }
-        }
-
-        // Compare region counts
-        if fp_a.regions.len() != fp_b.regions.len() {
-            differences.push(json!({
-                "field": "region_count",
-                "component_a": fp_a.regions.len(),
-                "component_b": fp_b.regions.len()
-            }));
-        }
-
-        // Compare text counts
-        if fp_a.text.len() != fp_b.text.len() {
-            differences.push(json!({
-                "field": "text_count",
-                "component_a": fp_a.text.len(),
-                "component_b": fp_b.text.len()
-            }));
-        }
-
-        // Compare 3D model references
-        let has_model_a = fp_a.model_3d.is_some();
-        let has_model_b = fp_b.model_3d.is_some();
-        if has_model_a != has_model_b {
-            differences.push(json!({
-                "field": "external_3d_model",
-                "component_a": has_model_a,
-                "component_b": has_model_b
-            }));
-        } else if has_model_a && has_model_b {
-            let m_a = fp_a.model_3d.as_ref().unwrap();
-            let m_b = fp_b.model_3d.as_ref().unwrap();
-            if m_a.filepath != m_b.filepath {
-                differences.push(json!({
-                    "field": "3d_model_path",
-                    "component_a": m_a.filepath,
-                    "component_b": m_b.filepath
-                }));
-            }
-        }
-
-        // Compare component bodies count
-        if fp_a.component_bodies.len() != fp_b.component_bodies.len() {
-            differences.push(json!({
-                "field": "component_body_count",
-                "component_a": fp_a.component_bodies.len(),
-                "component_b": fp_b.component_bodies.len()
-            }));
-        }
-
-        let is_identical = differences.is_empty();
-
-        let result = json!({
-            "status": "success",
-            "file_type": "PcbLib",
-            "component_a": {
-                "filepath": filepath_a,
-                "name": name_a
-            },
-            "component_b": {
-                "filepath": filepath_b,
-                "name": name_b
-            },
-            "identical": is_identical,
-            "difference_count": differences.len(),
-            "differences": differences,
-            "summary": {
-                "pads_a": fp_a.pads.len(),
-                "pads_b": fp_b.pads.len(),
-                "tracks_a": fp_a.tracks.len(),
-                "tracks_b": fp_b.tracks.len(),
-                "arcs_a": fp_a.arcs.len(),
-                "arcs_b": fp_b.arcs.len(),
-                "regions_a": fp_a.regions.len(),
-                "regions_b": fp_b.regions.len(),
-                "text_a": fp_a.text.len(),
-                "text_b": fp_b.text.len()
-            }
-        });
-
-        ToolCallResult::text(serde_json::to_string_pretty(&result).unwrap())
-    }
-
-    /// Compares two lists of pads.
-    fn compare_pads(
-        pads_a: &[crate::altium::pcblib::Pad],
-        pads_b: &[crate::altium::pcblib::Pad],
-        tolerance: f64,
-    ) -> Vec<Value> {
-        use std::collections::HashMap;
-
-        let mut diffs = Vec::new();
-
-        // Index pads by designator
-        let map_a: HashMap<&str, &crate::altium::pcblib::Pad> =
-            pads_a.iter().map(|p| (p.designator.as_str(), p)).collect();
-        let map_b: HashMap<&str, &crate::altium::pcblib::Pad> =
-            pads_b.iter().map(|p| (p.designator.as_str(), p)).collect();
-
-        // Find pads only in A
-        for des in map_a.keys() {
-            if !map_b.contains_key(des) {
-                diffs.push(json!({
-                    "designator": des,
-                    "status": "only_in_a"
-                }));
-            }
-        }
-
-        // Find pads only in B
-        for des in map_b.keys() {
-            if !map_a.contains_key(des) {
-                diffs.push(json!({
-                    "designator": des,
-                    "status": "only_in_b"
-                }));
-            }
-        }
-
-        // Compare matching pads
-        for (des, pad_a) in &map_a {
-            if let Some(pad_b) = map_b.get(des) {
-                let mut changes = Vec::new();
-
-                // Compare position
-                if (pad_a.x - pad_b.x).abs() > tolerance || (pad_a.y - pad_b.y).abs() > tolerance {
-                    changes.push(json!({
-                        "property": "position",
-                        "a": { "x": pad_a.x, "y": pad_a.y },
-                        "b": { "x": pad_b.x, "y": pad_b.y }
-                    }));
-                }
-
-                // Compare size
-                if (pad_a.width - pad_b.width).abs() > tolerance
-                    || (pad_a.height - pad_b.height).abs() > tolerance
-                {
-                    changes.push(json!({
-                        "property": "size",
-                        "a": { "width": pad_a.width, "height": pad_a.height },
-                        "b": { "width": pad_b.width, "height": pad_b.height }
-                    }));
-                }
-
-                // Compare shape
-                if pad_a.shape != pad_b.shape {
-                    changes.push(json!({
-                        "property": "shape",
-                        "a": format!("{:?}", pad_a.shape),
-                        "b": format!("{:?}", pad_b.shape)
-                    }));
-                }
-
-                // Compare layer
-                if pad_a.layer != pad_b.layer {
-                    changes.push(json!({
-                        "property": "layer",
-                        "a": pad_a.layer,
-                        "b": pad_b.layer
-                    }));
-                }
-
-                // Compare hole size
-                let hole_diff = match (pad_a.hole_size, pad_b.hole_size) {
-                    (Some(a), Some(b)) => (a - b).abs() > tolerance,
-                    (None, None) => false,
-                    _ => true,
-                };
-                if hole_diff {
-                    changes.push(json!({
-                        "property": "hole_size",
-                        "a": pad_a.hole_size,
-                        "b": pad_b.hole_size
-                    }));
-                }
-
-                // Compare rotation
-                if (pad_a.rotation - pad_b.rotation).abs() > tolerance {
-                    changes.push(json!({
-                        "property": "rotation",
-                        "a": pad_a.rotation,
-                        "b": pad_b.rotation
-                    }));
-                }
-
-                if !changes.is_empty() {
-                    diffs.push(json!({
-                        "designator": des,
-                        "status": "modified",
-                        "changes": changes
-                    }));
-                }
-            }
-        }
-
-        diffs
-    }
-
-    /// Compares two lists of tracks.
-    fn compare_tracks(
-        tracks_a: &[crate::altium::pcblib::Track],
-        tracks_b: &[crate::altium::pcblib::Track],
-        tolerance: f64,
-    ) -> Vec<Value> {
-        let mut diffs = Vec::new();
-
-        // For tracks, we compare by matching start/end coordinates
-        // Since tracks don't have unique identifiers, we'll report aggregate differences
-        let mut matched_b: Vec<bool> = vec![false; tracks_b.len()];
-
-        for (i, track_a) in tracks_a.iter().enumerate() {
-            let mut found_match = false;
-
-            for (j, track_b) in tracks_b.iter().enumerate() {
-                if matched_b[j] {
-                    continue;
-                }
-
-                // Check if tracks match (same endpoints within tolerance)
-                let same_forward = (track_a.x1 - track_b.x1).abs() <= tolerance
-                    && (track_a.y1 - track_b.y1).abs() <= tolerance
-                    && (track_a.x2 - track_b.x2).abs() <= tolerance
-                    && (track_a.y2 - track_b.y2).abs() <= tolerance;
-
-                let same_reverse = (track_a.x1 - track_b.x2).abs() <= tolerance
-                    && (track_a.y1 - track_b.y2).abs() <= tolerance
-                    && (track_a.x2 - track_b.x1).abs() <= tolerance
-                    && (track_a.y2 - track_b.y1).abs() <= tolerance;
-
-                if same_forward || same_reverse {
-                    matched_b[j] = true;
-                    found_match = true;
-
-                    // Check for width/layer differences
-                    let mut changes = Vec::new();
-                    if (track_a.width - track_b.width).abs() > tolerance {
-                        changes.push(json!({
-                            "property": "width",
-                            "a": track_a.width,
-                            "b": track_b.width
-                        }));
-                    }
-                    if track_a.layer != track_b.layer {
-                        changes.push(json!({
-                            "property": "layer",
-                            "a": track_a.layer,
-                            "b": track_b.layer
-                        }));
-                    }
-
-                    if !changes.is_empty() {
-                        diffs.push(json!({
-                            "track_index": i,
-                            "status": "modified",
-                            "endpoints": {
-                                "x1": track_a.x1, "y1": track_a.y1,
-                                "x2": track_a.x2, "y2": track_a.y2
-                            },
-                            "changes": changes
-                        }));
-                    }
-                    break;
-                }
-            }
-
-            if !found_match {
-                diffs.push(json!({
-                    "track_index": i,
-                    "status": "only_in_a",
-                    "endpoints": {
-                        "x1": track_a.x1, "y1": track_a.y1,
-                        "x2": track_a.x2, "y2": track_a.y2
-                    },
-                    "layer": track_a.layer,
-                    "width": track_a.width
-                }));
-            }
-        }
-
-        // Report unmatched tracks from B
-        for (j, track_b) in tracks_b.iter().enumerate() {
-            if !matched_b[j] {
-                diffs.push(json!({
-                    "track_index": j,
-                    "status": "only_in_b",
-                    "endpoints": {
-                        "x1": track_b.x1, "y1": track_b.y1,
-                        "x2": track_b.x2, "y2": track_b.y2
-                    },
-                    "layer": track_b.layer,
-                    "width": track_b.width
-                }));
-            }
-        }
-
-        diffs
-    }
-
-    /// Compares two lists of PCB arcs.
-    fn compare_pcb_arcs(
-        arcs_a: &[crate::altium::pcblib::Arc],
-        arcs_b: &[crate::altium::pcblib::Arc],
-        tolerance: f64,
-    ) -> Vec<Value> {
-        let mut diffs = Vec::new();
-        let mut matched_b: Vec<bool> = vec![false; arcs_b.len()];
-
-        for (i, arc_a) in arcs_a.iter().enumerate() {
-            let mut found_match = false;
-
-            for (j, arc_b) in arcs_b.iter().enumerate() {
-                if matched_b[j] {
-                    continue;
-                }
-
-                // Match by centre and radius
-                if (arc_a.x - arc_b.x).abs() <= tolerance
-                    && (arc_a.y - arc_b.y).abs() <= tolerance
-                    && (arc_a.radius - arc_b.radius).abs() <= tolerance
-                {
-                    matched_b[j] = true;
-                    found_match = true;
-
-                    let mut changes = Vec::new();
-                    if (arc_a.start_angle - arc_b.start_angle).abs() > tolerance {
-                        changes.push(json!({
-                            "property": "start_angle",
-                            "a": arc_a.start_angle,
-                            "b": arc_b.start_angle
-                        }));
-                    }
-                    if (arc_a.end_angle - arc_b.end_angle).abs() > tolerance {
-                        changes.push(json!({
-                            "property": "end_angle",
-                            "a": arc_a.end_angle,
-                            "b": arc_b.end_angle
-                        }));
-                    }
-                    if (arc_a.width - arc_b.width).abs() > tolerance {
-                        changes.push(json!({
-                            "property": "width",
-                            "a": arc_a.width,
-                            "b": arc_b.width
-                        }));
-                    }
-                    if arc_a.layer != arc_b.layer {
-                        changes.push(json!({
-                            "property": "layer",
-                            "a": arc_a.layer,
-                            "b": arc_b.layer
-                        }));
-                    }
-
-                    if !changes.is_empty() {
-                        diffs.push(json!({
-                            "arc_index": i,
-                            "status": "modified",
-                            "centre": { "x": arc_a.x, "y": arc_a.y },
-                            "radius": arc_a.radius,
-                            "changes": changes
-                        }));
-                    }
-                    break;
-                }
-            }
-
-            if !found_match {
-                diffs.push(json!({
-                    "arc_index": i,
-                    "status": "only_in_a",
-                    "centre": { "x": arc_a.x, "y": arc_a.y },
-                    "radius": arc_a.radius,
-                    "layer": arc_a.layer
-                }));
-            }
-        }
-
-        for (j, arc_b) in arcs_b.iter().enumerate() {
-            if !matched_b[j] {
-                diffs.push(json!({
-                    "arc_index": j,
-                    "status": "only_in_b",
-                    "centre": { "x": arc_b.x, "y": arc_b.y },
-                    "radius": arc_b.radius,
-                    "layer": arc_b.layer
-                }));
-            }
-        }
-
-        diffs
-    }
-
-    /// Compares two symbols in detail.
-    #[allow(clippy::too_many_lines)]
-    fn compare_symbols(
-        filepath_a: &str,
-        name_a: &str,
-        filepath_b: &str,
-        name_b: &str,
-        include_geometry: bool,
-    ) -> ToolCallResult {
-        use crate::altium::SchLib;
-
-        // Read libraries
-        let lib_a = match SchLib::open(filepath_a) {
-            Ok(lib) => lib,
-            Err(e) => return ToolCallResult::error(format!("Failed to read '{filepath_a}': {e}")),
-        };
-        let lib_b = match SchLib::open(filepath_b) {
-            Ok(lib) => lib,
-            Err(e) => return ToolCallResult::error(format!("Failed to read '{filepath_b}': {e}")),
-        };
-
-        // Get components
-        let Some(sym_a) = lib_a.get(name_a) else {
-            return ToolCallResult::error(format!(
-                "Component '{name_a}' not found in '{filepath_a}'"
-            ));
-        };
-        let Some(sym_b) = lib_b.get(name_b) else {
-            return ToolCallResult::error(format!(
-                "Component '{name_b}' not found in '{filepath_b}'"
-            ));
-        };
-
-        let mut differences: Vec<Value> = Vec::new();
-
-        // Compare description
-        if sym_a.description != sym_b.description {
-            differences.push(json!({
-                "field": "description",
-                "component_a": sym_a.description,
-                "component_b": sym_b.description
-            }));
-        }
-
-        // Compare designator
-        if sym_a.designator != sym_b.designator {
-            differences.push(json!({
-                "field": "designator",
-                "component_a": sym_a.designator,
-                "component_b": sym_b.designator
-            }));
-        }
-
-        // Compare pin counts
-        if sym_a.pins.len() != sym_b.pins.len() {
-            differences.push(json!({
-                "field": "pin_count",
-                "component_a": sym_a.pins.len(),
-                "component_b": sym_b.pins.len()
-            }));
-        }
-
-        // Compare pins in detail
-        if include_geometry {
-            let pin_diffs = Self::compare_pins(&sym_a.pins, &sym_b.pins);
-            if !pin_diffs.is_empty() {
-                differences.push(json!({
-                    "field": "pins",
-                    "differences": pin_diffs
-                }));
-            }
-        }
-
-        // Compare rectangle counts
-        if sym_a.rectangles.len() != sym_b.rectangles.len() {
-            differences.push(json!({
-                "field": "rectangle_count",
-                "component_a": sym_a.rectangles.len(),
-                "component_b": sym_b.rectangles.len()
-            }));
-        }
-
-        // Compare line counts
-        if sym_a.lines.len() != sym_b.lines.len() {
-            differences.push(json!({
-                "field": "line_count",
-                "component_a": sym_a.lines.len(),
-                "component_b": sym_b.lines.len()
-            }));
-        }
-
-        // Compare polyline counts
-        if sym_a.polylines.len() != sym_b.polylines.len() {
-            differences.push(json!({
-                "field": "polyline_count",
-                "component_a": sym_a.polylines.len(),
-                "component_b": sym_b.polylines.len()
-            }));
-        }
-
-        // Compare arc counts
-        if sym_a.arcs.len() != sym_b.arcs.len() {
-            differences.push(json!({
-                "field": "arc_count",
-                "component_a": sym_a.arcs.len(),
-                "component_b": sym_b.arcs.len()
-            }));
-        }
-
-        // Compare footprint references
-        if sym_a.footprints.len() != sym_b.footprints.len() {
-            differences.push(json!({
-                "field": "footprint_count",
-                "component_a": sym_a.footprints.len(),
-                "component_b": sym_b.footprints.len()
-            }));
-        }
-
-        // Compare footprint names
-        if include_geometry {
-            let fps_a: std::collections::HashSet<&str> =
-                sym_a.footprints.iter().map(|f| f.name.as_str()).collect();
-            let fps_b: std::collections::HashSet<&str> =
-                sym_b.footprints.iter().map(|f| f.name.as_str()).collect();
-
-            let only_in_a: Vec<_> = fps_a.difference(&fps_b).copied().collect();
-            let only_in_b: Vec<_> = fps_b.difference(&fps_a).copied().collect();
-
-            if !only_in_a.is_empty() || !only_in_b.is_empty() {
-                differences.push(json!({
-                    "field": "footprints",
-                    "only_in_a": only_in_a,
-                    "only_in_b": only_in_b
-                }));
-            }
-        }
-
-        // Compare parameters
-        if include_geometry {
-            let param_diffs = Self::compare_parameters(&sym_a.parameters, &sym_b.parameters);
-            if !param_diffs.is_empty() {
-                differences.push(json!({
-                    "field": "parameters",
-                    "differences": param_diffs
-                }));
-            }
-        }
-
-        let is_identical = differences.is_empty();
-
-        let result = json!({
-            "status": "success",
-            "file_type": "SchLib",
-            "component_a": {
-                "filepath": filepath_a,
-                "name": name_a
-            },
-            "component_b": {
-                "filepath": filepath_b,
-                "name": name_b
-            },
-            "identical": is_identical,
-            "difference_count": differences.len(),
-            "differences": differences,
-            "summary": {
-                "pins_a": sym_a.pins.len(),
-                "pins_b": sym_b.pins.len(),
-                "rectangles_a": sym_a.rectangles.len(),
-                "rectangles_b": sym_b.rectangles.len(),
-                "lines_a": sym_a.lines.len(),
-                "lines_b": sym_b.lines.len(),
-                "parameters_a": sym_a.parameters.len(),
-                "parameters_b": sym_b.parameters.len(),
-                "footprints_a": sym_a.footprints.len(),
-                "footprints_b": sym_b.footprints.len()
-            }
-        });
-
-        ToolCallResult::text(serde_json::to_string_pretty(&result).unwrap())
-    }
-
-    /// Compares two lists of pins.
-    fn compare_pins(
-        pins_a: &[crate::altium::schlib::Pin],
-        pins_b: &[crate::altium::schlib::Pin],
-    ) -> Vec<Value> {
-        use std::collections::HashMap;
-
-        let mut diffs = Vec::new();
-
-        // Index pins by designator
-        let map_a: HashMap<&str, &crate::altium::schlib::Pin> =
-            pins_a.iter().map(|p| (p.designator.as_str(), p)).collect();
-        let map_b: HashMap<&str, &crate::altium::schlib::Pin> =
-            pins_b.iter().map(|p| (p.designator.as_str(), p)).collect();
-
-        // Find pins only in A
-        for des in map_a.keys() {
-            if !map_b.contains_key(des) {
-                diffs.push(json!({
-                    "designator": des,
-                    "status": "only_in_a"
-                }));
-            }
-        }
-
-        // Find pins only in B
-        for des in map_b.keys() {
-            if !map_a.contains_key(des) {
-                diffs.push(json!({
-                    "designator": des,
-                    "status": "only_in_b"
-                }));
-            }
-        }
-
-        // Compare matching pins
-        for (des, pin_a) in &map_a {
-            if let Some(pin_b) = map_b.get(des) {
-                let mut changes = Vec::new();
-
-                // Compare position
-                if pin_a.x != pin_b.x || pin_a.y != pin_b.y {
-                    changes.push(json!({
-                        "property": "position",
-                        "a": { "x": pin_a.x, "y": pin_a.y },
-                        "b": { "x": pin_b.x, "y": pin_b.y }
-                    }));
-                }
-
-                // Compare length
-                if pin_a.length != pin_b.length {
-                    changes.push(json!({
-                        "property": "length",
-                        "a": pin_a.length,
-                        "b": pin_b.length
-                    }));
-                }
-
-                // Compare name
-                if pin_a.name != pin_b.name {
-                    changes.push(json!({
-                        "property": "name",
-                        "a": pin_a.name,
-                        "b": pin_b.name
-                    }));
-                }
-
-                // Compare electrical type
-                if pin_a.electrical_type != pin_b.electrical_type {
-                    changes.push(json!({
-                        "property": "electrical_type",
-                        "a": format!("{:?}", pin_a.electrical_type),
-                        "b": format!("{:?}", pin_b.electrical_type)
-                    }));
-                }
-
-                // Compare orientation
-                if pin_a.orientation != pin_b.orientation {
-                    changes.push(json!({
-                        "property": "orientation",
-                        "a": format!("{:?}", pin_a.orientation),
-                        "b": format!("{:?}", pin_b.orientation)
-                    }));
-                }
-
-                if !changes.is_empty() {
-                    diffs.push(json!({
-                        "designator": des,
-                        "status": "modified",
-                        "changes": changes
-                    }));
-                }
-            }
-        }
-
-        diffs
-    }
-
-    /// Compares two lists of parameters.
-    fn compare_parameters(
-        params_a: &[crate::altium::schlib::Parameter],
-        params_b: &[crate::altium::schlib::Parameter],
-    ) -> Vec<Value> {
-        use std::collections::HashMap;
-
-        let mut diffs = Vec::new();
-
-        // Index parameters by name
-        let map_a: HashMap<&str, &crate::altium::schlib::Parameter> =
-            params_a.iter().map(|p| (p.name.as_str(), p)).collect();
-        let map_b: HashMap<&str, &crate::altium::schlib::Parameter> =
-            params_b.iter().map(|p| (p.name.as_str(), p)).collect();
-
-        // Find parameters only in A
-        for (name, param) in &map_a {
-            if !map_b.contains_key(name) {
-                diffs.push(json!({
-                    "name": name,
-                    "status": "only_in_a",
-                    "value": param.value
-                }));
-            }
-        }
-
-        // Find parameters only in B
-        for (name, param) in &map_b {
-            if !map_a.contains_key(name) {
-                diffs.push(json!({
-                    "name": name,
-                    "status": "only_in_b",
-                    "value": param.value
-                }));
-            }
-        }
-
-        // Compare matching parameters
-        for (name, param_a) in &map_a {
-            if let Some(param_b) = map_b.get(name) {
-                if param_a.value != param_b.value {
-                    diffs.push(json!({
-                        "name": name,
-                        "status": "modified",
-                        "value_a": param_a.value,
-                        "value_b": param_b.value
-                    }));
-                }
-            }
-        }
-
-        diffs
     }
 
     /// Performs batch updates across all components in a library file.
@@ -8755,738 +7079,6 @@ impl McpServer {
         });
 
         ToolCallResult::text(serde_json::to_string_pretty(&result).unwrap())
-    }
-
-    // ==================== Rendering Tools ====================
-
-    /// Renders an ASCII art visualisation of a footprint.
-    #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-    fn call_render_footprint(&self, arguments: &Value) -> ToolCallResult {
-        use crate::altium::PcbLib;
-
-        let Some(filepath) = arguments.get("filepath").and_then(Value::as_str) else {
-            return ToolCallResult::error("Missing required parameter: filepath");
-        };
-
-        let Some(component_name) = arguments.get("component_name").and_then(Value::as_str) else {
-            return ToolCallResult::error("Missing required parameter: component_name");
-        };
-
-        // Validate path is within allowed directories
-        if let Err(e) = self.validate_path(filepath) {
-            return ToolCallResult::error(e);
-        }
-
-        // Parse optional parameters
-        let scale = arguments
-            .get("scale")
-            .and_then(Value::as_f64)
-            .unwrap_or(2.0);
-        let max_width = arguments
-            .get("max_width")
-            .and_then(Value::as_u64)
-            .unwrap_or(80) as usize;
-        let max_height = arguments
-            .get("max_height")
-            .and_then(Value::as_u64)
-            .unwrap_or(40) as usize;
-
-        if scale <= 0.0 {
-            return ToolCallResult::error("scale must be greater than 0");
-        }
-
-        // Read the library
-        let library = match PcbLib::open(filepath) {
-            Ok(lib) => lib,
-            Err(e) => return ToolCallResult::error(format!("Failed to read library: {e}")),
-        };
-
-        // Find the footprint
-        let Some(footprint) = library.get(component_name) else {
-            let available: Vec<_> = library.iter().take(5).map(|f| &f.name).collect();
-            let hint = if available.is_empty() {
-                "Library is empty".to_string()
-            } else {
-                format!(
-                    "Available footprints include: {}{}",
-                    available
-                        .iter()
-                        .map(|s| s.as_str())
-                        .collect::<Vec<_>>()
-                        .join(", "),
-                    if library.len() > 5 {
-                        format!(" (and {} more)", library.len() - 5)
-                    } else {
-                        String::new()
-                    }
-                )
-            };
-            return ToolCallResult::error(format!(
-                "Footprint '{component_name}' not found in library. {hint}"
-            ));
-        };
-
-        // Render the footprint
-        let ascii_art = Self::render_footprint_ascii(footprint, scale, max_width, max_height);
-
-        let result = json!({
-            "status": "success",
-            "filepath": filepath,
-            "component_name": component_name,
-            "scale": scale,
-            "render": ascii_art,
-        });
-
-        ToolCallResult::text(serde_json::to_string_pretty(&result).unwrap())
-    }
-
-    /// Renders a footprint as ASCII art.
-    #[allow(
-        clippy::cast_possible_truncation,
-        clippy::cast_sign_loss,
-        clippy::cast_precision_loss,
-        clippy::similar_names,
-        clippy::too_many_lines,
-        clippy::float_cmp,
-        clippy::needless_range_loop
-    )]
-    fn render_footprint_ascii(
-        footprint: &crate::altium::pcblib::Footprint,
-        scale: f64,
-        max_width: usize,
-        max_height: usize,
-    ) -> String {
-        use std::fmt::Write;
-
-        // Find bounding box
-        let (mut min_x, mut max_x, mut min_y, mut max_y) = (f64::MAX, f64::MIN, f64::MAX, f64::MIN);
-
-        for pad in &footprint.pads {
-            let half_w = pad.width / 2.0;
-            let half_h = pad.height / 2.0;
-            min_x = min_x.min(pad.x - half_w);
-            max_x = max_x.max(pad.x + half_w);
-            min_y = min_y.min(pad.y - half_h);
-            max_y = max_y.max(pad.y + half_h);
-        }
-
-        for track in &footprint.tracks {
-            min_x = min_x.min(track.x1.min(track.x2) - track.width / 2.0);
-            max_x = max_x.max(track.x1.max(track.x2) + track.width / 2.0);
-            min_y = min_y.min(track.y1.min(track.y2) - track.width / 2.0);
-            max_y = max_y.max(track.y1.max(track.y2) + track.width / 2.0);
-        }
-
-        for arc in &footprint.arcs {
-            min_x = min_x.min(arc.x - arc.radius);
-            max_x = max_x.max(arc.x + arc.radius);
-            min_y = min_y.min(arc.y - arc.radius);
-            max_y = max_y.max(arc.y + arc.radius);
-        }
-
-        // Handle empty footprint
-        if min_x == f64::MAX {
-            return "Empty footprint (no primitives)".to_string();
-        }
-
-        // Add margin
-        let margin = 0.5;
-        min_x -= margin;
-        max_x += margin;
-        min_y -= margin;
-        max_y += margin;
-
-        // Calculate canvas size
-        let width_mm = max_x - min_x;
-        let height_mm = max_y - min_y;
-        let mut canvas_width = (width_mm * scale).ceil() as usize;
-        let mut canvas_height = (height_mm * scale).ceil() as usize;
-
-        // Clamp to max dimensions
-        if canvas_width > max_width {
-            canvas_width = max_width;
-        }
-        if canvas_height > max_height {
-            canvas_height = max_height;
-        }
-
-        // Ensure minimum size
-        canvas_width = canvas_width.max(10);
-        canvas_height = canvas_height.max(5);
-
-        // Calculate actual scale after clamping
-        let actual_scale_x = canvas_width as f64 / width_mm;
-        let actual_scale_y = canvas_height as f64 / height_mm;
-
-        // Create canvas (y is inverted for display)
-        let mut canvas = vec![vec![' '; canvas_width]; canvas_height];
-
-        // Helper to convert coordinates
-        let to_canvas = |x: f64, y: f64| -> (usize, usize) {
-            let cx = ((x - min_x) * actual_scale_x).round() as usize;
-            let cy =
-                canvas_height.saturating_sub(1) - ((y - min_y) * actual_scale_y).round() as usize;
-            (cx.min(canvas_width - 1), cy.min(canvas_height - 1))
-        };
-
-        // Draw tracks (as lines)
-        for track in &footprint.tracks {
-            Self::draw_line(
-                &mut canvas,
-                to_canvas(track.x1, track.y1),
-                to_canvas(track.x2, track.y2),
-                '-',
-            );
-        }
-
-        // Draw arcs (simplified as circles at centre)
-        for arc in &footprint.arcs {
-            let (cx, cy) = to_canvas(arc.x, arc.y);
-            if cx < canvas_width && cy < canvas_height {
-                canvas[cy][cx] = 'o';
-            }
-        }
-
-        // Draw pads (as rectangles with full designator)
-        for pad in &footprint.pads {
-            let half_w = pad.width / 2.0;
-            let half_h = pad.height / 2.0;
-            let (x1, y1) = to_canvas(pad.x - half_w, pad.y - half_h);
-            let (x2, y2) = to_canvas(pad.x + half_w, pad.y + half_h);
-
-            // Fill pad area
-            let (min_cy, max_cy) = (y1.min(y2), y1.max(y2));
-            let (min_cx, max_cx) = (x1.min(x2), x1.max(x2));
-
-            for cy in min_cy..=max_cy {
-                for cx in min_cx..=max_cx {
-                    if cy < canvas_height && cx < canvas_width {
-                        canvas[cy][cx] = '#';
-                    }
-                }
-            }
-
-            // Place full designator centred on pad
-            let (cx, cy) = to_canvas(pad.x, pad.y);
-            if cy < canvas_height {
-                let desig = &pad.designator;
-                let start_x = cx.saturating_sub(desig.len() / 2);
-                for (i, ch) in desig.chars().enumerate() {
-                    let x = start_x + i;
-                    if x < canvas_width {
-                        canvas[cy][x] = ch;
-                    }
-                }
-            }
-        }
-
-        // Draw origin crosshair
-        let (ox, oy) = to_canvas(0.0, 0.0);
-        if ox < canvas_width && oy < canvas_height {
-            canvas[oy][ox] = '+';
-        }
-
-        // Build output string
-        let mut output = String::new();
-        let _ = writeln!(
-            output,
-            "Footprint: {} ({:.2} x {:.2} mm)",
-            footprint.name,
-            width_mm - margin * 2.0,
-            height_mm - margin * 2.0
-        );
-        let _ = writeln!(
-            output,
-            "Pads: {}, Tracks: {}, Arcs: {}",
-            footprint.pads.len(),
-            footprint.tracks.len(),
-            footprint.arcs.len()
-        );
-        output.push_str(&"-".repeat(canvas_width + 2));
-        output.push('\n');
-
-        for row in &canvas {
-            output.push('|');
-            for &ch in row {
-                output.push(ch);
-            }
-            output.push('|');
-            output.push('\n');
-        }
-
-        output.push_str(&"-".repeat(canvas_width + 2));
-        output.push('\n');
-        output.push_str("Legend: # = pad, - = track, o = arc, + = origin\n");
-
-        output
-    }
-
-    /// Draws a line on the canvas using Bresenham's algorithm.
-    #[allow(clippy::cast_possible_wrap, clippy::cast_sign_loss)]
-    fn draw_line(
-        canvas: &mut [Vec<char>],
-        (x0, y0): (usize, usize),
-        (x1, y1): (usize, usize),
-        ch: char,
-    ) {
-        let dx = (x1 as isize - x0 as isize).abs();
-        let dy = (y1 as isize - y0 as isize).abs();
-        let sx: isize = if x0 < x1 { 1 } else { -1 };
-        let sy: isize = if y0 < y1 { 1 } else { -1 };
-        let mut err = dx - dy;
-
-        let mut x = x0 as isize;
-        let mut y = y0 as isize;
-
-        let height = canvas.len();
-        let width = if height > 0 { canvas[0].len() } else { 0 };
-
-        loop {
-            if (x as usize) < width && (y as usize) < height {
-                canvas[y as usize][x as usize] = ch;
-            }
-
-            if x == x1 as isize && y == y1 as isize {
-                break;
-            }
-
-            let e2 = 2 * err;
-            if e2 > -dy {
-                err -= dy;
-                x += sx;
-            }
-            if e2 < dx {
-                err += dx;
-                y += sy;
-            }
-        }
-    }
-
-    /// Renders an ASCII art visualisation of a schematic symbol.
-    #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-    fn call_render_symbol(&self, arguments: &Value) -> ToolCallResult {
-        use crate::altium::SchLib;
-
-        let Some(filepath) = arguments.get("filepath").and_then(Value::as_str) else {
-            return ToolCallResult::error("Missing required parameter: filepath");
-        };
-
-        let Some(component_name) = arguments.get("component_name").and_then(Value::as_str) else {
-            return ToolCallResult::error("Missing required parameter: component_name");
-        };
-
-        // Validate path is within allowed directories
-        if let Err(e) = self.validate_path(filepath) {
-            return ToolCallResult::error(e);
-        }
-
-        // Parse optional parameters
-        let scale = arguments
-            .get("scale")
-            .and_then(Value::as_f64)
-            .unwrap_or(1.0);
-        let max_width = arguments
-            .get("max_width")
-            .and_then(Value::as_u64)
-            .unwrap_or(80) as usize;
-        let max_height = arguments
-            .get("max_height")
-            .and_then(Value::as_u64)
-            .unwrap_or(40) as usize;
-        let part_id = arguments
-            .get("part_id")
-            .and_then(Value::as_i64)
-            .unwrap_or(1) as i32;
-
-        if scale <= 0.0 {
-            return ToolCallResult::error("scale must be greater than 0");
-        }
-
-        // Read the library
-        let library = match SchLib::open(filepath) {
-            Ok(lib) => lib,
-            Err(e) => return ToolCallResult::error(format!("Failed to read library: {e}")),
-        };
-
-        // Find the symbol
-        let Some(symbol) = library.get(component_name) else {
-            let available: Vec<_> = library.iter().take(5).map(|s| s.name.as_str()).collect();
-            let hint = if available.is_empty() {
-                "Library is empty".to_string()
-            } else {
-                format!(
-                    "Available symbols include: {}{}",
-                    available.join(", "),
-                    if library.len() > 5 {
-                        format!(" (and {} more)", library.len() - 5)
-                    } else {
-                        String::new()
-                    }
-                )
-            };
-            return ToolCallResult::error(format!(
-                "Symbol '{component_name}' not found in library. {hint}"
-            ));
-        };
-
-        // Render the symbol
-        let ascii_art = Self::render_symbol_ascii(symbol, scale, max_width, max_height, part_id);
-
-        let result = json!({
-            "status": "success",
-            "filepath": filepath,
-            "component_name": component_name,
-            "scale": scale,
-            "part_id": part_id,
-            "render": ascii_art,
-        });
-
-        ToolCallResult::text(serde_json::to_string_pretty(&result).unwrap())
-    }
-
-    /// Renders a schematic symbol as ASCII art.
-    #[allow(
-        clippy::cast_possible_truncation,
-        clippy::cast_sign_loss,
-        clippy::cast_precision_loss,
-        clippy::similar_names,
-        clippy::too_many_lines,
-        clippy::float_cmp,
-        clippy::needless_range_loop
-    )]
-    fn render_symbol_ascii(
-        symbol: &crate::altium::schlib::Symbol,
-        scale: f64,
-        max_width: usize,
-        max_height: usize,
-        part_id: i32,
-    ) -> String {
-        use crate::altium::schlib::PinOrientation;
-        use std::fmt::Write;
-
-        // Find bounding box (in schematic units)
-        let (mut min_x, mut max_x, mut min_y, mut max_y) = (i32::MAX, i32::MIN, i32::MAX, i32::MIN);
-
-        // Helper to check if primitive belongs to requested part
-        let matches_part = |owner_part_id: i32| -> bool {
-            part_id == 0 || owner_part_id == part_id || owner_part_id == 0
-        };
-
-        // Calculate bounding box from pins (include pin length)
-        for pin in &symbol.pins {
-            if !matches_part(pin.owner_part_id) {
-                continue;
-            }
-            let (px, py) = (pin.x, pin.y);
-            let (end_x, end_y) = match pin.orientation {
-                PinOrientation::Right => (px + pin.length, py),
-                PinOrientation::Left => (px - pin.length, py),
-                PinOrientation::Up => (px, py + pin.length),
-                PinOrientation::Down => (px, py - pin.length),
-            };
-            min_x = min_x.min(px).min(end_x);
-            max_x = max_x.max(px).max(end_x);
-            min_y = min_y.min(py).min(end_y);
-            max_y = max_y.max(py).max(end_y);
-        }
-
-        // Calculate bounding box from rectangles
-        for rect in &symbol.rectangles {
-            if !matches_part(rect.owner_part_id) {
-                continue;
-            }
-            min_x = min_x.min(rect.x1).min(rect.x2);
-            max_x = max_x.max(rect.x1).max(rect.x2);
-            min_y = min_y.min(rect.y1).min(rect.y2);
-            max_y = max_y.max(rect.y1).max(rect.y2);
-        }
-
-        // Calculate bounding box from lines
-        for line in &symbol.lines {
-            if !matches_part(line.owner_part_id) {
-                continue;
-            }
-            min_x = min_x.min(line.x1).min(line.x2);
-            max_x = max_x.max(line.x1).max(line.x2);
-            min_y = min_y.min(line.y1).min(line.y2);
-            max_y = max_y.max(line.y1).max(line.y2);
-        }
-
-        // Calculate bounding box from polylines
-        for polyline in &symbol.polylines {
-            if !matches_part(polyline.owner_part_id) {
-                continue;
-            }
-            for &(x, y) in &polyline.points {
-                min_x = min_x.min(x);
-                max_x = max_x.max(x);
-                min_y = min_y.min(y);
-                max_y = max_y.max(y);
-            }
-        }
-
-        // Calculate bounding box from arcs
-        for arc in &symbol.arcs {
-            if !matches_part(arc.owner_part_id) {
-                continue;
-            }
-            min_x = min_x.min(arc.x - arc.radius);
-            max_x = max_x.max(arc.x + arc.radius);
-            min_y = min_y.min(arc.y - arc.radius);
-            max_y = max_y.max(arc.y + arc.radius);
-        }
-
-        // Calculate bounding box from ellipses
-        for ellipse in &symbol.ellipses {
-            if !matches_part(ellipse.owner_part_id) {
-                continue;
-            }
-            min_x = min_x.min(ellipse.x - ellipse.radius_x);
-            max_x = max_x.max(ellipse.x + ellipse.radius_x);
-            min_y = min_y.min(ellipse.y - ellipse.radius_y);
-            max_y = max_y.max(ellipse.y + ellipse.radius_y);
-        }
-
-        // Handle empty symbol
-        if min_x == i32::MAX {
-            return "Empty symbol (no primitives)".to_string();
-        }
-
-        // Add margin (10 schematic units = 1 grid)
-        let margin = 10;
-        min_x -= margin;
-        max_x += margin;
-        min_y -= margin;
-        max_y += margin;
-
-        // Calculate canvas size (scale is chars per 10 schematic units)
-        let width_units = f64::from(max_x - min_x);
-        let height_units = f64::from(max_y - min_y);
-        let mut canvas_width = ((width_units / 10.0) * scale).ceil() as usize;
-        let mut canvas_height = ((height_units / 10.0) * scale).ceil() as usize;
-
-        // Clamp to max dimensions
-        canvas_width = canvas_width.clamp(10, max_width);
-        canvas_height = canvas_height.clamp(5, max_height);
-
-        // Calculate actual scale after clamping
-        let actual_scale_x = canvas_width as f64 / width_units;
-        let actual_scale_y = canvas_height as f64 / height_units;
-
-        // Create canvas (y is inverted for display)
-        let mut canvas = vec![vec![' '; canvas_width]; canvas_height];
-
-        // Helper to convert schematic coordinates to canvas coordinates
-        let to_canvas = |x: i32, y: i32| -> (usize, usize) {
-            let cx = (f64::from(x - min_x) * actual_scale_x).round() as usize;
-            let cy = canvas_height.saturating_sub(1)
-                - (f64::from(y - min_y) * actual_scale_y).round() as usize;
-            (cx.min(canvas_width - 1), cy.min(canvas_height - 1))
-        };
-
-        // Draw rectangles (as box outlines or filled)
-        for rect in &symbol.rectangles {
-            if !matches_part(rect.owner_part_id) {
-                continue;
-            }
-            let (x1, y1) = to_canvas(rect.x1, rect.y1);
-            let (x2, y2) = to_canvas(rect.x2, rect.y2);
-            let (min_cx, max_cx) = (x1.min(x2), x1.max(x2));
-            let (min_cy, max_cy) = (y1.min(y2), y1.max(y2));
-
-            // Draw top and bottom edges
-            for cx in min_cx..=max_cx {
-                if cx < canvas_width {
-                    if min_cy < canvas_height {
-                        canvas[min_cy][cx] = '-';
-                    }
-                    if max_cy < canvas_height {
-                        canvas[max_cy][cx] = '-';
-                    }
-                }
-            }
-            // Draw left and right edges
-            for cy in min_cy..=max_cy {
-                if cy < canvas_height {
-                    if min_cx < canvas_width {
-                        canvas[cy][min_cx] = '|';
-                    }
-                    if max_cx < canvas_width {
-                        canvas[cy][max_cx] = '|';
-                    }
-                }
-            }
-            // Draw corners
-            if min_cy < canvas_height && min_cx < canvas_width {
-                canvas[min_cy][min_cx] = '+';
-            }
-            if min_cy < canvas_height && max_cx < canvas_width {
-                canvas[min_cy][max_cx] = '+';
-            }
-            if max_cy < canvas_height && min_cx < canvas_width {
-                canvas[max_cy][min_cx] = '+';
-            }
-            if max_cy < canvas_height && max_cx < canvas_width {
-                canvas[max_cy][max_cx] = '+';
-            }
-        }
-
-        // Draw lines
-        for line in &symbol.lines {
-            if !matches_part(line.owner_part_id) {
-                continue;
-            }
-            Self::draw_line(
-                &mut canvas,
-                to_canvas(line.x1, line.y1),
-                to_canvas(line.x2, line.y2),
-                '-',
-            );
-        }
-
-        // Draw polylines
-        for polyline in &symbol.polylines {
-            if !matches_part(polyline.owner_part_id) {
-                continue;
-            }
-            for i in 0..polyline.points.len().saturating_sub(1) {
-                let (x1, y1) = polyline.points[i];
-                let (x2, y2) = polyline.points[i + 1];
-                Self::draw_line(&mut canvas, to_canvas(x1, y1), to_canvas(x2, y2), '-');
-            }
-        }
-
-        // Draw arcs (simplified as circles at centre)
-        for arc in &symbol.arcs {
-            if !matches_part(arc.owner_part_id) {
-                continue;
-            }
-            let (cx, cy) = to_canvas(arc.x, arc.y);
-            if cx < canvas_width && cy < canvas_height {
-                canvas[cy][cx] = 'o';
-            }
-        }
-
-        // Draw ellipses (simplified as circles at centre)
-        for ellipse in &symbol.ellipses {
-            if !matches_part(ellipse.owner_part_id) {
-                continue;
-            }
-            let (cx, cy) = to_canvas(ellipse.x, ellipse.y);
-            if cx < canvas_width && cy < canvas_height {
-                canvas[cy][cx] = 'O';
-            }
-        }
-
-        // Draw pins with full designators
-        for pin in &symbol.pins {
-            if !matches_part(pin.owner_part_id) {
-                continue;
-            }
-            let (px, py) = (pin.x, pin.y);
-            let (end_x, end_y) = match pin.orientation {
-                PinOrientation::Right => (px + pin.length, py),
-                PinOrientation::Left => (px - pin.length, py),
-                PinOrientation::Up => (px, py + pin.length),
-                PinOrientation::Down => (px, py - pin.length),
-            };
-
-            // Draw pin line
-            Self::draw_line(&mut canvas, to_canvas(px, py), to_canvas(end_x, end_y), '~');
-
-            // Draw full designator at connection point
-            let (cx, cy) = to_canvas(px, py);
-            if cy < canvas_height {
-                let desig = &pin.designator;
-
-                // Place designator based on pin orientation
-                match pin.orientation {
-                    PinOrientation::Right => {
-                        // Pin points right, place designator to the left (inside symbol)
-                        for (i, ch) in desig.chars().enumerate() {
-                            let x = cx.saturating_sub(desig.len() - 1 - i);
-                            if x < canvas_width {
-                                canvas[cy][x] = ch;
-                            }
-                        }
-                    }
-                    PinOrientation::Left => {
-                        // Pin points left, place designator to the right (inside symbol)
-                        for (i, ch) in desig.chars().enumerate() {
-                            let x = cx + i;
-                            if x < canvas_width {
-                                canvas[cy][x] = ch;
-                            }
-                        }
-                    }
-                    PinOrientation::Up | PinOrientation::Down => {
-                        // Vertical pins: place designator horizontally centred
-                        let start_x = cx.saturating_sub(desig.len() / 2);
-                        for (i, ch) in desig.chars().enumerate() {
-                            let x = start_x + i;
-                            if x < canvas_width {
-                                canvas[cy][x] = ch;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        // Draw origin crosshair
-        let (ox, oy) = to_canvas(0, 0);
-        if ox < canvas_width && oy < canvas_height && canvas[oy][ox] == ' ' {
-            canvas[oy][ox] = '+';
-        }
-
-        // Count primitives for summary
-        let pin_count = symbol
-            .pins
-            .iter()
-            .filter(|p| matches_part(p.owner_part_id))
-            .count();
-        let rect_count = symbol
-            .rectangles
-            .iter()
-            .filter(|r| matches_part(r.owner_part_id))
-            .count();
-        let line_count = symbol
-            .lines
-            .iter()
-            .filter(|l| matches_part(l.owner_part_id))
-            .count();
-
-        // Build output string
-        let mut output = String::new();
-        let _ = writeln!(
-            output,
-            "Symbol: {} (part {}/{})",
-            symbol.name,
-            if part_id == 0 { 1 } else { part_id },
-            symbol.part_count
-        );
-        let _ = writeln!(
-            output,
-            "Pins: {pin_count}, Rectangles: {rect_count}, Lines: {line_count}"
-        );
-        output.push_str(&"-".repeat(canvas_width + 2));
-        output.push('\n');
-
-        for row in &canvas {
-            output.push('|');
-            for &ch in row {
-                output.push(ch);
-            }
-            output.push('|');
-            output.push('\n');
-        }
-
-        output.push_str(&"-".repeat(canvas_width + 2));
-        output.push('\n');
-        output
-            .push_str("Legend: |-+ = rectangle, ~ = pin line, o = arc, O = ellipse, + = origin\n");
-
-        output
     }
 
     /// Manages parameters in a `SchLib` symbol.
